+++ site/components/menu/.package/DBUpdates.php
@@ -13,7 +13,8 @@ class DBUpdates extends \components\update\classes\BaseDBUpdates
       '1.1' => '1.2',
       '1.2' => '1.3',
       '1.3' => '1.4',
-      '1.4' => '0.2.0-beta' //No DB changes.
+      '1.4' => '0.2.0-beta', //No DB changes.
+      '0.2.0-beta' => '0.3.0-beta' //No DB changes.
     );
   
   //Add new installer to prevent pagetype problems.